修改bug

+++ Data_structure/Lab1/InToSuf/Stack.h
@@ -8,7 +8,7 @@
 typedef struct intstack
 {
     int top;
-    int data[MAXLENGTH];
+    float data[MAXLENGTH];
 } Istack;
 
 typedef struct charstack
@@ -17,12 +17,12 @@ typedef struct charstack
     char data[MAXLENGTH];
 } Cstack;
 
-void ICreat(Istack *s);             //创建空栈
-void IMakNull(Istack *s);           //将栈置空
-int ITop(Istack *s);                //返回栈顶值
-void IPop(Istack *s);               //弹栈
-void IPush(Istack *s, int element); //压栈
-void IShowStack(Istack *s);         //显示栈内容，从栈底开始
+void ICreat(Istack *s);               //创建空栈
+void IMakNull(Istack *s);             //将栈置空
+float ITop(Istack *s);                //返回栈顶值
+void IPop(Istack *s);                 //弹栈
+void IPush(Istack *s, float element); //压栈
+void IShowStack(Istack *s);           //显示栈内容，从栈底开始
 
 void CCreat(Cstack *s);              //创建空栈
 void CMakeNull(Cstack *s);           //将栈置空
